Moved client wrappers to linera-service/src/client.rs
