+++ linera-service/src/lib.rs
@@ -5,6 +5,7 @@
 //! This module provides the executables needed to operate a Linera service, including a placeholder wallet acting as a GraphQL service for user interfaces.
 
 pub mod chain_listener;
+pub mod client;
 pub mod config;
 pub mod grpc_proxy;
 pub mod node_service;
